Revert "Split driver and rider flow"
