+++ app/src/main/java/com/example/suberduberuber/Activities/DashboardActivity.java
@@ -2,6 +2,7 @@ package com.example.suberduberuber.Activities;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.ActionBarDrawerToggle;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
 import androidx.appcompat.widget.Toolbar;
@@ -17,15 +18,17 @@ import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
 
+import com.example.suberduberuber.Fragments.SelectDestinationFragment;
 import com.example.suberduberuber.R;
 import com.google.android.material.navigation.NavigationView;
 import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
 
-abstract class DashboardActivity extends AppCompatActivity {
+public class DashboardActivity extends AppCompatActivity {
 
     private FirebaseAuth myAuth;
 
-    protected NavController navController;
+    private NavController navController;
 
     private DrawerLayout drawerLayout;
     private NavigationView navigationView;
@@ -33,11 +36,11 @@ abstract class DashboardActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(getContentViewId());
+        setContentView(R.layout.activity_dashboard);
 
         myAuth = FirebaseAuth.getInstance();
 
-        navController = Navigation.findNavController(this, getNavHostId());
+        navController = Navigation.findNavController(this, R.id.nav_host_fragment);
 
         navigationView = findViewById(R.id.nav_view);
 
@@ -51,13 +54,10 @@ abstract class DashboardActivity extends AppCompatActivity {
         configureNavigationDrawer();
     }
 
-    abstract int getNavHostId();
-    abstract int getContentViewId();
-
     // for QR Code Fragment
     public void onActivityResult(int requestCode, int resultCode, Intent intent) {
         super.onActivityResult(requestCode, resultCode, intent);
-        Fragment fragment = getSupportFragmentManager().findFragmentById(getNavHostId());
+        Fragment fragment = getSupportFragmentManager().findFragmentById(R.id.nav_host_fragment);
         if (fragment != null) {
             fragment.onActivityResult(requestCode, resultCode, intent);
         }
