+++ app/src/main/java/com/example/suberduberuber/ViewModels/AuthViewModel.java
@@ -1,10 +1,8 @@
 package com.example.suberduberuber.ViewModels;
 
 import android.app.Application;
-import android.util.Log;
 
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import androidx.lifecycle.AndroidViewModel;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
@@ -15,13 +13,8 @@ import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.firestore.DocumentSnapshot;
-import com.google.firebase.firestore.EventListener;
-import com.google.firebase.firestore.FirebaseFirestoreException;
-import com.google.firebase.firestore.QuerySnapshot;
 
 public class AuthViewModel extends AndroidViewModel {
-
-    private static final String TAG = "AUTH_VIEW_MODEL";
 
     private UserRepository userRepository;
 
@@ -30,23 +23,5 @@ public class AuthViewModel extends AndroidViewModel {
     public AuthViewModel(Application application) {
         super(application);
         userRepository = new UserRepository();
-    }
-
-    // Get the current user and setup listening for live updates
-    public LiveData<User> getCurrentUser() {
-        userRepository.getCurrentUser().addSnapshotListener(new EventListener<QuerySnapshot>() {
-            @Override
-            public void onEvent(@Nullable QuerySnapshot queryDocumentSnapshots, @Nullable FirebaseFirestoreException e) {
-                if(e != null || queryDocumentSnapshots.getDocuments().isEmpty()) {
-                    currentUser.setValue(null);
-                    Log.e(TAG, "Could not get user.", e);
-                }
-                else {
-                    currentUser.setValue(queryDocumentSnapshots.getDocuments().get(0).toObject(User.class));
-                }
-            }
-        });
-
-        return currentUser;
     }
 }
