--- app/src/main/java/com/example/suberduberuber/ViewRequestsFragment.java
@@ -1,28 +1,0 @@
-package com.example.suberduberuber;
-
-import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-
-/**
- * A simple {@link Fragment} subclass.
- */
-public class ViewRequestsFragment extends Fragment {
-
-    public ViewRequestsFragment() {
-        // Required empty public constructor
-    }
-
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_view_requests, container, false);
-    }
-}
