--- app/src/main/res/navigation/driver_nav_graph.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/driver_nav_graph"
-    app:startDestination="@id/viewRequestsFragment">
-
-    <fragment
-        android:id="@+id/viewRequestsFragment"
-        android:name="com.example.suberduberuber.ViewRequestsFragment"
-        android:label="fragment_view_requests"
-        tools:layout="@layout/fragment_view_requests" />
-</navigation>
